+++ relay-dynamic-config/src/feature.rs
@@ -19,6 +19,10 @@ pub enum Feature {
     /// Enables metric extraction from spans.
     #[serde(rename = "projects:span-metrics-extraction")]
     SpanMetricsExtraction,
+    /// Extract spans from transactions and convert them to standalone spans.
+    #[serde(rename = "projects:extract-standalone-spans")]
+    ExtractStandaloneSpans,
+
     /// Deprecated, still forwarded for older downstream Relays.
     #[serde(rename = "organizations:transaction-name-mark-scrubbed-as-sanitized")]
     Deprecated1,
@@ -28,7 +32,6 @@ pub enum Feature {
     /// Deprecated, still forwarded for older downstream Relays.
     #[serde(rename = "organizations:profiling")]
     Deprecated3,
-
     /// Forward compatibility.
     #[serde(other)]
     Unknown,