+++ relay-sampling/src/lib.rs
@@ -2388,7 +2388,7 @@ mod tests {
         let rules: Result<Vec<RuleCondition>, _> = serde_json::from_str(serialized_rules);
         assert!(rules.is_ok());
         let rules = rules.unwrap();
-        insta::assert_ron_snapshot!(rules, @r###"
+        insta::assert_ron_snapshot!(rules, @r#"
             [
               EqCondition(
                 op: "eq",
@@ -2451,7 +2451,7 @@ mod tests {
                   ),
                 ],
               ),
-            ]"###);
+            ]"#);
     }
 
     #[test]
@@ -2794,13 +2794,13 @@ mod tests {
             Ok(MatchedRuleIds(vec![RuleId(123)]))
         );
 
-        assert!(matches!(MatchedRuleIds::from_string(""), Err(_)));
+        assert!(MatchedRuleIds::from_string("").is_err());
 
-        assert!(matches!(MatchedRuleIds::from_string(","), Err(_)));
+        assert!(MatchedRuleIds::from_string(",").is_err());
 
-        assert!(matches!(MatchedRuleIds::from_string("123.456"), Err(_)));
+        assert!(MatchedRuleIds::from_string("123.456").is_err());
 
-        assert!(matches!(MatchedRuleIds::from_string("a,b"), Err(_)));
+        assert!(MatchedRuleIds::from_string("a,b").is_err());
     }
 
     #[test]
@@ -3207,7 +3207,7 @@ mod tests {
         }
         "#;
         let dsc = serde_json::from_str::<DynamicSamplingContext>(json).unwrap();
-        insta::assert_ron_snapshot!(dsc, @r###"
+        insta::assert_ron_snapshot!(dsc, @r#"
         {
           "trace_id": "00000000-0000-0000-0000-000000000000",
           "public_key": "abd0f232775f45feab79864e580d160b",
@@ -3217,7 +3217,7 @@ mod tests {
           "user_id": "hello",
           "replay_id": None,
         }
-        "###);
+        "#);
     }
 
     #[test]
@@ -3231,7 +3231,7 @@ mod tests {
         }
         "#;
         let dsc = serde_json::from_str::<DynamicSamplingContext>(json).unwrap();
-        insta::assert_ron_snapshot!(dsc, @r###"
+        insta::assert_ron_snapshot!(dsc, @r#"
         {
           "trace_id": "00000000-0000-0000-0000-000000000000",
           "public_key": "abd0f232775f45feab79864e580d160b",
@@ -3242,7 +3242,7 @@ mod tests {
           "user_id": "hello",
           "replay_id": None,
         }
-        "###);
+        "#);
     }
 
     #[test]
@@ -3256,7 +3256,7 @@ mod tests {
         }
         "#;
         let dsc = serde_json::from_str::<DynamicSamplingContext>(json).unwrap();
-        insta::assert_ron_snapshot!(dsc, @r###"
+        insta::assert_ron_snapshot!(dsc, @r#"
         {
           "trace_id": "00000000-0000-0000-0000-000000000000",
           "public_key": "abd0f232775f45feab79864e580d160b",
@@ -3267,7 +3267,7 @@ mod tests {
           "user_id": "hello",
           "replay_id": None,
         }
-        "###);
+        "#);
     }
 
     #[test]
@@ -3321,7 +3321,7 @@ mod tests {
         "#;
         let dsc = serde_json::from_str::<DynamicSamplingContext>(json).unwrap();
         let dsc_as_json = serde_json::to_string_pretty(&dsc).unwrap();
-        let expected_json = r###"{
+        let expected_json = r#"{
   "trace_id": "00000000-0000-0000-0000-000000000000",
   "public_key": "abd0f232775f45feab79864e580d160b",
   "release": null,
@@ -3330,7 +3330,7 @@ mod tests {
   "user_id": "hello",
   "replay_id": null,
   "sampled": true
-}"###;
+}"#;
 
         assert_eq!(dsc_as_json, expected_json);
     }
@@ -3347,7 +3347,7 @@ mod tests {
         "#;
         let dsc = serde_json::from_str::<DynamicSamplingContext>(json).unwrap();
         let dsc_as_json = serde_json::to_string_pretty(&dsc).unwrap();
-        let expected_json = r###"{
+        let expected_json = r#"{
   "trace_id": "00000000-0000-0000-0000-000000000000",
   "public_key": "abd0f232775f45feab79864e580d160b",
   "release": null,
@@ -3356,7 +3356,7 @@ mod tests {
   "user_id": "hello",
   "replay_id": null,
   "sampled": false
-}"###;
+}"#;
 
         assert_eq!(dsc_as_json, expected_json);
     }
@@ -3373,7 +3373,7 @@ mod tests {
         "#;
         let dsc = serde_json::from_str::<DynamicSamplingContext>(json).unwrap();
         let dsc_as_json = serde_json::to_string_pretty(&dsc).unwrap();
-        let expected_json = r###"{
+        let expected_json = r#"{
   "trace_id": "00000000-0000-0000-0000-000000000000",
   "public_key": "abd0f232775f45feab79864e580d160b",
   "release": null,
@@ -3381,7 +3381,7 @@ mod tests {
   "transaction": null,
   "user_id": "hello",
   "replay_id": null
-}"###;
+}"#;
 
         assert_eq!(dsc_as_json, expected_json);
     }
@@ -3398,7 +3398,7 @@ mod tests {
         "#;
         let dsc = serde_json::from_str::<DynamicSamplingContext>(json).unwrap();
         let dsc_as_json = serde_json::to_string_pretty(&dsc).unwrap();
-        let expected_json = r###"{
+        let expected_json = r#"{
   "trace_id": "00000000-0000-0000-0000-000000000000",
   "public_key": "abd0f232775f45feab79864e580d160b",
   "release": null,
@@ -3406,7 +3406,7 @@ mod tests {
   "transaction": null,
   "user_id": "hello",
   "replay_id": null
-}"###;
+}"#;
 
         assert_eq!(dsc_as_json, expected_json);
     }