+++ relay-server/src/metrics_extraction/generic.rs
@@ -177,7 +177,7 @@ mod tests {
         let config = serde_json::from_value(config_json).unwrap();
 
         let metrics = extract_metrics_from(event.value().unwrap(), &config);
-        insta::assert_debug_snapshot!(metrics, @r###"
+        insta::assert_debug_snapshot!(metrics, @r#"
         [
             Metric {
                 name: "c:transactions/counter@none",
@@ -188,7 +188,7 @@ mod tests {
                 tags: {},
             },
         ]
-        "###);
+        "#);
     }
 
     #[test]
@@ -213,7 +213,7 @@ mod tests {
         let config = serde_json::from_value(config_json).unwrap();
 
         let metrics = extract_metrics_from(event.value().unwrap(), &config);
-        insta::assert_debug_snapshot!(metrics, @r###"
+        insta::assert_debug_snapshot!(metrics, @r#"
         [
             Metric {
                 name: "d:transactions/duration@none",
@@ -224,7 +224,7 @@ mod tests {
                 tags: {},
             },
         ]
-        "###);
+        "#);
     }
 
     #[test]
@@ -251,7 +251,7 @@ mod tests {
         let config = serde_json::from_value(config_json).unwrap();
 
         let metrics = extract_metrics_from(event.value().unwrap(), &config);
-        insta::assert_debug_snapshot!(metrics, @r###"
+        insta::assert_debug_snapshot!(metrics, @r#"
         [
             Metric {
                 name: "s:transactions/users@none",
@@ -262,7 +262,7 @@ mod tests {
                 tags: {},
             },
         ]
-        "###);
+        "#);
     }
 
     #[test]
@@ -301,7 +301,7 @@ mod tests {
         let config = serde_json::from_value(config_json).unwrap();
 
         let metrics = extract_metrics_from(event.value().unwrap(), &config);
-        insta::assert_debug_snapshot!(metrics, @r###"
+        insta::assert_debug_snapshot!(metrics, @r#"
         [
             Metric {
                 name: "c:transactions/counter@none",
@@ -316,6 +316,6 @@ mod tests {
                 },
             },
         ]
-        "###);
+        "#);
     }
 }